Mtrk.3592508 added the following.: - app/Http/Controllers/Auth/AuthController.php - app/Http/Controllers/Controller.php - nohup.out - resources/views/Auth/login.blade.php - resources/views/Auth/registration.blade.php - resources/views/dashboard.blade.php - routes/web.php
This implements a functional Login Sequence by going to http://85.215.157.91/login
As of right now, Login is verified by email and password.
The registration Process currently DOES NOT send an Email to a User creating an account.

The nohup file is everchanging due to tracking interaction with the php artisan webserver.
This is NOT good logging and isnt supposed to be. The only intended purpose is to have it run in the background
so its not in the way actively and to allow reading it for debugging.